% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cata_ryebread.R
\docType{data}
\name{cata_ryebread}
\alias{cata_ryebread}
\title{cata_ryebread data}
\format{
CATA data. A data frame with 6 rows (the number of ryebread) and 1848 columns (the number of consumers (132) * the number of attributes (14)).
 For each consumer,each attribute and each product, there is 1 if the attribute has been checked by the consumer for the product, and 0 if not.
}
\usage{
data(cata_ryebread)
}
\description{
cata_ryebread data
}
\examples{
data(cata_ryebread)
}
\references{
Giacalone, D. (2018). Product Performance Optimization. In Ares, G., & Varela, P. (Eds.)
Methods in Consumer Research, Volume I (Chapter 7. pp. 159-185), Elsevier.
}
\keyword{datasets}
