% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheese.R
\docType{data}
\name{cheese}
\alias{cheese}
\title{cheese Just About Right data}
\format{
JAR data. A data frame with Assessors, Products and JAR attributes. 8 products, 9 attributes and 72 subjects.
}
\usage{
data(cheese)
}
\description{
cheese Just About Right data
}
\examples{
data(cheese)
}
\references{
Luc, A., Lê, S., Philippe, M., Qannari, E. M., & Vigneau, E. (2022). Free JAR experiment: Data analysis and comparison with JAR task. Food Quality and Preference, 98, 104453.
}
\keyword{datasets}
