% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choc.R
\docType{data}
\name{choc}
\alias{choc}
\title{chocolates data}
\format{
Free sorting data. A data frame with 14 rows (the chocolates) and 25 columns (the subjects).
 The numbers indicate the groups to which the products (rows) are assigned.
}
\usage{
data(choc)
}
\description{
chocolates data
}
\examples{
data(choc)
}
\references{
Courcoux, P., Qannari, E. M., Taylor, Y., Buck, D., & Greenhoff, K. (2012). Taxonomic free
sorting. Food Quality and Preference, 23(1), 30-35.
}
\keyword{datasets}
