% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluscata_liking.R
\name{cluscata_liking}
\alias{cluscata_liking}
\title{Perform a cluster analysis of subjects on CATA/liking combination}
\usage{
cluscata_liking(Data, nblo, NameBlocks=NULL, NameVar=NULL,  Itermax=30,
                       Graph_dend=TRUE, Graph_bar=TRUE,
                        printlevel=FALSE,gpmax=min(6, nblo-1))
}
\arguments{
\item{Data}{data frame or matrix where the blocks of variables (attributes) are merged horizontally thanks to \code{\link{combinCATALiking}} function}

\item{nblo}{numerical. Number of blocks (subjects).}

\item{NameBlocks}{string vector. Name of each block (subject). Length must be equal to the number of blocks. If NULL, the names are S1,...Sm. Default: NULL}

\item{NameVar}{string vector. Name of each variable (attribute, the same names for each subject). Length must be equal to the number of attributes. If NULL, the colnames of the first block are taken. Default: NULL}

\item{Itermax}{numerical. Maximum of iteration for the partitioning algorithm. Default:30}

\item{Graph_dend}{logical. Should the dendrogram be plotted? Default: TRUE}

\item{Graph_bar}{logical. Should the barplot of the difference of the criterion and the barplot of the overall homogeneity at each merging step of the hierarchical algorithm be plotted? Default: TRUE}

\item{printlevel}{logical. Print the number of remaining levels during the hierarchical clustering algorithm? Default: FALSE}

\item{gpmax}{logical. What is maximum number of clusters to consider? Default: min(6, nblo-2)}
}
\value{
Each partitionK contains a list for each number of clusters of the partition, K=1 to gpmax with:
        \itemize{
         \item group: the clustering partition after consolidation.
         \item compromise: the compromise of each cluster
         \item dist_all_cluster: the distance between each subject and each cluster compromise
         \item criterion: the CLUSCATA-liking criterion error
         \item param: parameters called
         \item type: parameter passed to other functions
         }
         There is also at the end of the list:
         \itemize{
         \item dend: The CLUSCATA dendrogram
         \item cutree_k: the partition obtained by cutting the dendrogram in K clusters (before consolidation).
         \item diff_crit_ng: variation of criterion when a merging is done before consolidation (and after)
         \item param: parameters called
         \item type: parameter passed to other functions
         }
}
\description{
Clustering of subjects (blocks) from combination of CATA and liking experiments.
}
\examples{
data(cata_ryebread)
data(liking_ryebread)
cataliking=combinCATALiking(cata_ryebread, liking_ryebread)

#with only 40 subjects
resclustcatal=cluscata_liking(Data=cataliking[,1:(40*14)], nblo=40, gpmax=5)
plot(resclustcatal, cata_ryebread[,1:(40*14)], liking_ryebread[,1:40])


}
\references{
Vigneau, E., Cariou, V., Giacalone, D., Berget, I., & Llobell, F. (2022). Combining hedonic information and CATA description for consumer segmentation. Food Quality and Preference, 95, 104358.
}
\seealso{
\code{\link{plot.cluscata_liking}}, \code{\link{summary.cluscata_liking}} , \code{\link{combinCATALiking}}
}
\keyword{CATA}
\keyword{liking}
