% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinCATALiking.R
\name{combinCATALiking}
\alias{combinCATALiking}
\title{Combination of CATA and liking data for CLUSCATA-liking}
\usage{
combinCATALiking(cata, liking, center=TRUE, scale=FALSE)
}
\arguments{
\item{cata}{data frame or matrix where the blocks of binary variables are merged horizontally. If you have a different format, see \code{\link{change_cata_format}}}

\item{liking}{data frame or matrix where the products are in rows and the assessors in columns}

\item{center}{Centering of consumer liking. Default: TRUE}

\item{scale}{Scaling  of consumer liking. Default: FALSE}
}
\value{
Combined data
}
\description{
For CLUSCATA-liking, this preprocessing is needed.
}
\examples{
data(cata_ryebread)
data(liking_ryebread)
cataliking=combinCATALiking(cata_ryebread, liking_ryebread)

}
\references{
Vigneau, E., Cariou, V., Giacalone, D., Berget, I., & Llobell, F. (2022). Combining hedonic information and CATA description for consumer segmentation. Food Quality and Preference, 95, 104358.
}
\seealso{
\code{\link{cluscata_liking}}
}
\keyword{CATA}
\keyword{liking}
