% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/croissant.R
\docType{data}
\name{croissant}
\alias{croissant}
\title{Croissant JAR and liking data}
\format{
Just-About-Right (JAR) and liking data. A data frame with one row
per subject-product combination. The first column identifies the subjects,
the second column identifies the products, the following columns contain
the JAR evaluations on a 5-point scale, and the last column contains the
overall liking scores.
}
\usage{
data(croissant)
}
\description{
Croissant JAR and liking data
}
\examples{
data(croissant)
}
\references{
Llobell, F. & Guksch, T. (2026). Beyond penalty analysis:
Joint clustering of consumers using JAR and liking data.
EuroSense, Oslo, Norway.
}
\keyword{datasets}
