% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish.R
\docType{data}
\name{fish}
\alias{fish}
\title{fish data}
\format{
CATA data with sessions. A data frame with the sessions, the panelists, the products and CATA attributes.
}
\usage{
data(fish)
}
\description{
fish data
}
\examples{
data(fish)
}
\references{
Bonnet, L., Ferney, T., Riedel, T., Qannari, E.M., Llobell, F. (September 14, 2022) .Using CATA for sensory profiling: assessment of the panel performance. Eurosense, Turku, Finland.
}
\keyword{datasets}
