% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liking_ryebread.R
\docType{data}
\name{liking_ryebread}
\alias{liking_ryebread}
\title{liking_ryebread data}
\format{
Liking data. A data frame with 6 rows (the number of ryebread) and 132 columns (the number of consumers)
}
\usage{
data(liking_ryebread)
}
\description{
liking_ryebread data
}
\examples{
data(liking_ryebread)
}
\references{
Giacalone, D. (2018). Product Performance Optimization. In Ares, G., & Varela, P. (Eds.)
Methods in Consumer Research, Volume I (Chapter 7. pp. 159-185), Elsevier.
}
\keyword{datasets}
