% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixclustatis.R
\name{mixclustatis}
\alias{mixclustatis}
\title{Perform a cluster analysis of variables}
\usage{
mixclustatis(Data, quanti=NULL,quali=NULL,Noise_cluster=FALSE,
Itermax=20, printlevel=FALSE, Graph_dend=TRUE, Graph_bar=TRUE,
gpmax=min(6, length(quali)+length(quanti)-1), rhoparam = NULL)
}
\arguments{
\item{Data}{data frame or matrix. Correspond to all the data (variables are columns)}

\item{quanti}{numerical vector. The number of the columns containing quantitative variables.}

\item{quali}{numerical vector. The number of the columns containing qualitative variables.}

\item{Noise_cluster}{logical. Should a noise cluster be computed? Default: FALSE}

\item{Itermax}{numerical. Maximum of iteration for the partitioning algorithm. Default: 30}

\item{printlevel}{logical. Print the number of remaining levels during the hierarchical clustering algorithm? Default: FALSE}

\item{Graph_dend}{logical. Should the dendrogram be plotted? Default: TRUE}

\item{Graph_bar}{logical. Should the barplot of the difference of the criterion and the barplot of the overall homogeneity at each merging step of the hierarchical algorithm be plotted? Default: TRUE}

\item{gpmax}{logical. What is maximum number of clusters to consider? Default:  min(6, number of variables -2)}

\item{rhoparam}{numerical or vector. What is the threshold for the noise cluster? Between 0 and 1, high value can imply lot of vatriables set aside. If NULL, automatic threshold is computed. Can be different for each group (in this case, provide a vector)}
}
\value{
Each partitionK contains a list for each number of clusters of the partition, K=1 to gpmax with:
        \itemize{
         \item group: the clustering partition of variables after consolidation. If Noise_cluster=TRUE, some variables could be in the noise cluster ("K+1")
         \item rho: the threshold(s) for the noise cluster (computed or input parameter)
         \item homogeneity: homogeneity index (%) of each cluster and the overall homogeneity index (%) of the partition
         \item rv_with_compromise: RV coefficient of each variable with its cluster compromise
         \item weights: weight associated with each variable in its cluster
         \item comp_RV: RV coefficient between the compromises associated with the various clusters
         \item compromise: the W compromise of each cluster
         \item coord: the coordinates of objects of each cluster
         \item inertia: percentage of total variance explained by each axis for each cluster
         \item rv_all_cluster: the RV coefficient between each variable and each cluster compromise
         \item criterion: the CLUSTATIS criterion error
         \item param: parameters called in the consolidation
         \item type: parameter passed to other functions
         }
         There is also at the end of the list:
         \itemize{
         \item dend: The CLUSTATIS dendrogram
         \item cutree_k: the partition obtained by cutting the dendrogram for K clusters (before consolidation).
         \item overall_homogeneity_ng: percentage of overall homogeneity by number of clusters before consolidation (and after if there is no noise cluster)
         \item diff_crit_ng: variation of criterion when a merging is done before consolidation (and after if there is no noise cluster)
         \item param: parameters called
         \item type: parameter passed to other functions
         }
}
\description{
Perform cluster analysis of variables in context of quantitative,
 qualitative or mixed datasets with MixCluStatis.
}
\examples{
data("wine", package = "FactoMineR")
res=mixclustatis(wine, quanti = 3:29, quali = 1:2)
summary(res)
plot(res, Graph_groups = FALSE, Graph_weights = TRUE)

}
\references{
Paper submitted: Llobell, F., Abdi, H., Eslami, A. (2026). Clustering of categorical and mixed data variables around latent variables.
Llobell, F., Cariou, V., Vigneau, E., Labenne, A., & Qannari, E. M. (2018). Analysis and clustering of multiblock datasets by means of the STATIS and CLUSTATIS methods. Application to sensometrics. Food Quality and Preference, in Press.\cr
Llobell, F., Vigneau, E., Qannari, E. M. (2019). Clustering datasets by means of CLUSTATIS with identification of atypical datasets. Application to sensometrics. Food Quality and Preference, 75, 97-104.
Llobell, F., & Qannari, E. M. (2020). CLUSTATIS: Cluster analysis of blocks of variables. Electronic Journal of Applied Statistical Analysis, 13(2).
}
\seealso{
\code{\link{clustatis}}, \code{\link{plot.clustatis}}, \code{\link{summary.clustatis}}
}
\keyword{variables}
