% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PB_tuna.R
\docType{data}
\name{PB_tuna}
\alias{PB_tuna}
\title{Plant based tuna data}
\format{CATA data with sessions. A data frame with the sessions, the panelists, the products and CATA attributes.}
\usage{
data(PB_tuna)
}
\description{
Plant based tuna data
}
\examples{
data(PB_tuna)
}
\references{
Bonnet, L., Ferney, T., Riedel, T., Qannari, E.M., Llobell, F. (September 14, 2022) .Using CATA for sensory profiling: assessment of the panel performance. Eurosense, Turku, Finland.
}
\keyword{datasets}
