% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cluscata_liking.R
\name{plot.cluscata_liking}
\alias{plot.cluscata_liking}
\title{Displays the CLUSCATA-liking graphs}
\usage{
\method{plot}{cluscata_liking}(x, DataCata, Dataliking, ngroups=NULL,
 center=TRUE, scale=FALSE,
 sep_graphs=TRUE, Graph_dend=TRUE, Graph_bar=FALSE,
  cex=1,  xlimfreq=c(0,0.6), ylimchangelik=c(-2, 2), ...)
}
\arguments{
\item{x}{object of class 'cluscata_liking'.}

\item{DataCata}{data frame or matrix where the blocks of binary variables are merged horizontally. If you have a different format, see \code{\link{change_cata_format}}}

\item{Dataliking}{data frame or matrix where the products are in rows and the assessors in columns}

\item{ngroups}{number of groups to consider. Default: recommended  number of clusters}

\item{center}{Centering of consumer liking. Default: TRUE}

\item{scale}{Scaling  of consumer liking. Default: FALSE}

\item{sep_graphs}{logical. Should cata and liking data analyzed separately first?}

\item{Graph_dend}{logical. Should the dendrogram be plotted? Default: TRUE}

\item{Graph_bar}{logical. Should the barplot of the difference of the criterion and the barplot of the overall homogeneity at each merging step of the hierarchical algorithm be plotted? Also available after consolidation if Noise_cluster=FALSE. Default: FALSE}

\item{cex}{numerical. Numeric character expansion factor; multiplied by par("cex") yields the final character size. NULL and NA are equivalent to 1.0.}

\item{xlimfreq}{vector of 2 values. Used for the graph of the frequency of citations}

\item{ylimchangelik}{vector of 2 values. Used for the graphs of change in liking}

\item{...}{further arguments passed to or from other methods}
}
\value{
the CLUSCATA-liking graphs
}
\description{
This function plots dendrogram, variation of the merging criterion, weights and CATATIS map of each cluster
}
\examples{
\donttest{
data(cata_ryebread)
data(liking_ryebread)
cataliking=combinCATALiking(cata_ryebread, liking_ryebread)

#with only 40 subjects
resclustcatal=cluscata_liking(Data=cataliking[,1:(40*14)], nblo=40, gpmax=5)
plot(resclustcatal, cata_ryebread[,1:(40*14)], liking_ryebread[,1:40],
xlimfreq=c(0,0.6),  ylimchangelik=c(-2, 2))

}

}
\seealso{
\code{\link{cluscata_liking}} , \code{\link{summary.cluscata_liking}}
}
\keyword{CATA}
\keyword{liking}
