% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.clusrows.R
\name{plot.clusRows}
\alias{plot.clusRows}
\title{Displays the ClusMB and clustRowsOnstatisAxes graphs}
\usage{
\method{plot}{clusRows}(x, ...)
}
\arguments{
\item{x}{object of class 'clusRows'}

\item{...}{further arguments passed to or from other methods}
}
\value{
the dendrogram
}
\description{
This function plots the dendrogram of ClusMB or clustRowsOnstatisAxes
}
\examples{
##'
#####projective mapping####
library(ClustBlock)
data(smoo)
res1=ClusMB(smoo, rep(2,24))
plot(res1)

}
\seealso{
\code{\link{ClusMB}}, \code{\link{clustRowsOnStatisAxes}}
}
\keyword{CATA}
\keyword{RATA}
\keyword{quantitative}
