% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.catatis.R
\name{print.catatis}
\alias{print.catatis}
\title{Print the CATATIS results}
\usage{
\method{print}{catatis}(x, ...)
}
\arguments{
\item{x}{object of class 'catatis'}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print the CATATIS results
}
\seealso{
\code{\link{catatis}}
}
\keyword{CATA}
\keyword{RATA}
