% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cluscata.R
\name{print.cluscata}
\alias{print.cluscata}
\title{Print the CLUSCATA results}
\usage{
\method{print}{cluscata}(x, ...)
}
\arguments{
\item{x}{object of class 'cluscata'}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print the CLUSCATA results
}
\seealso{
\code{\link{cluscata}} , \code{\link{cluscata_kmeans}}
}
\keyword{CATA}
