% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cluscata_liking.R
\name{print.cluscata_liking}
\alias{print.cluscata_liking}
\title{Print the CLUSCATA-liking results}
\usage{
\method{print}{cluscata_liking}(x, ...)
}
\arguments{
\item{x}{object of class 'cluscata_liking'}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print the CLUSCATA-liking results
}
\seealso{
\code{\link{cluscata_liking}}
}
\keyword{CATA}
\keyword{liking}
