% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.clusRows.R
\name{print.clusRows}
\alias{print.clusRows}
\title{Print the ClusMB or clustering on STATIS axes results}
\usage{
\method{print}{clusRows}(x, ...)
}
\arguments{
\item{x}{object of class 'clusRows'}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print the ClusMB or clustering on STATIS axes results
}
\seealso{
\code{\link{ClusMB}}, \code{\link{clustRowsOnStatisAxes}}
}
\keyword{CATA}
\keyword{RATA}
\keyword{quantitative}
