% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.clustatis.R
\name{print.clustatis}
\alias{print.clustatis}
\title{Print the CLUSTATIS results}
\usage{
\method{print}{clustatis}(x, ...)
}
\arguments{
\item{x}{object of class 'clustatis'}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print the CLUSTATIS results
}
\seealso{
\code{\link{clustatis}}  , \code{\link{clustatis_kmeans}}
}
\keyword{quantitative}
