% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.statis.R
\name{print.statis}
\alias{print.statis}
\title{Print the STATIS results}
\usage{
\method{print}{statis}(x, ...)
}
\arguments{
\item{x}{object of class 'statis'}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print the STATIS results
}
\seealso{
\code{\link{statis}}
}
\keyword{quantitative}
