% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoo.R
\docType{data}
\name{smoo}
\alias{smoo}
\title{smoothies data}
\format{
Projective mapping (or Napping) data. A data frame with 8 rows (the number of smoothies) and 48 columns (the number of consumers * 2).
 For each consumer, we have the coordinates of the products on the sheet of paper.
}
\usage{
data(smoo)
}
\description{
smoothies data
}
\examples{
data(smoo)
}
\references{
Francois Husson, Sebastien Le and Marine Cadoret (2017). SensoMineR: Sensory Data Analysis. R
package version 1.23. https://CRAN.R-project.org/package=SensoMineR
}
\keyword{datasets}
