% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/straw.R
\docType{data}
\name{straw}
\alias{straw}
\title{strawberries data}
\format{
CATA data. A data frame with 6 rows (the number of strawberries) and 1824 columns (the number of consumers (114) * the number of attributes (16)).
 For each consumer,each attribute and eachb product, there is 1 if the attribute has been checked by the consumer for the product, and 0 if not.
}
\usage{
data(straw)
}
\description{
strawberries data
}
\examples{
data(straw)
}
\references{
Ares, G., & Jaeger, S. R. (2013). Check-all-that-apply questions: Influence of attribute
order on sensory product characterization. Food Quality and Preference, 28(1), 141-153.
}
\keyword{datasets}
