% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.catatis.R
\name{summary.catatis}
\alias{summary.catatis}
\title{Show the CATATIS results}
\usage{
\method{summary}{catatis}(object, ...)
}
\arguments{
\item{object}{object of class 'catatis'.}

\item{...}{further arguments passed to or from other methods}
}
\value{
a list with:
        \itemize{
         \item homogeneity: homogeneity of the subjects (in percentage)
         \item weights: the weights associated with the subjects to build the compromise
         \item eigenvalues: the eigenvalues associated to the correspondance analysis
         \item inertia: the percentage of total variance explained by each axis of the CA
         }
}
\description{
This function shows the CATATIS results
}
\seealso{
\code{\link{catatis}}
}
\keyword{CATA}
\keyword{RATA}
