% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.cluscata.R
\name{summary.cluscata}
\alias{summary.cluscata}
\title{Show the CLUSCATA results}
\usage{
\method{summary}{cluscata}(object, ngroups=NULL, ...)
}
\arguments{
\item{object}{object of class 'cluscata'.}

\item{ngroups}{number of groups to consider. Ignored for cluscata_kmeans results. Default: recommended  number of clusters}

\item{...}{further arguments passed to or from other methods}
}
\value{
the CLUSCATA principal results

a list with:
        \itemize{
         \item group: the clustering partition
         \item homogeneity: homogeneity index (%) of each cluster and the overall homogeneity index (%) of the partition
         \item weights: weight associated with each subject in its cluster
         \item rho: the threshold for the noise cluster
         \item test_one_cluster: decision and pvalue to know if there is more than one cluster
         }
}
\description{
This function shows the cluscata results
}
\seealso{
\code{\link{cluscata}} , \code{\link{cluscata_kmeans}}
}
\keyword{CATA}
