% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.cluscata_liking.R
\name{summary.cluscata_liking}
\alias{summary.cluscata_liking}
\title{Show the CLUSCATA-liking results}
\usage{
\method{summary}{cluscata_liking}(object, ngroups=NULL, ...)
}
\arguments{
\item{object}{object of class 'cluscata_liking'.}

\item{ngroups}{number of groups to consider. Default: recommended  number of clusters}

\item{...}{further arguments passed to or from other methods}
}
\value{
the CLUSCATA-liking principal results

a list with:
        \itemize{
         \item group: the clustering partition
         \item groupsvector: the groups in a vector
         }
}
\description{
This function shows the cluscata_liking results
}
\seealso{
\code{\link{cluscata_liking}} , \code{\link{plot.cluscata_liking}}
}
\keyword{CATA}
\keyword{liking}
