% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.clusRows.R
\name{summary.clusRows}
\alias{summary.clusRows}
\title{Show the ClusMB or clustering on STATIS axes results}
\usage{
\method{summary}{clusRows}(object, ...)
}
\arguments{
\item{object}{object of class 'clusRows'.}

\item{...}{further arguments passed to or from other methods}
}
\value{
a list with:
        \itemize{
         \item groups: clustering partition
         \item nbClustRetained: the number of clusters retained
         \item nbgH: Advised number of clusters per Hartigan index
         \item nbgCH: Advised number of clusters per Calinski-Harabasz index
         }
}
\description{
This function shows the ClusMB or clustering on STATIS axes results
}
\seealso{
\code{\link{ClusMB}}, \code{\link{clustRowsOnStatisAxes}}
}
\keyword{CATA}
\keyword{RATA}
\keyword{quantitative}
