% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.clustatis.R
\name{summary.clustatis}
\alias{summary.clustatis}
\title{Show the CLUSTATIS results}
\usage{
\method{summary}{clustatis}(object, ngroups=NULL, ...)
}
\arguments{
\item{object}{object of class 'clustatis'.}

\item{ngroups}{number of groups to consider. Ignored for clustatis_kmeans results. Default: recommended  number of clusters}

\item{...}{further arguments passed to or from other methods}
}
\value{
the CLUSTATIS principal results

a list with:
        \itemize{
         \item group: the clustering partition
         \item homogeneity: homogeneity index (%) of each cluster and the overall homogeneity index (%) of the partition
         \item weights: weight associated with each block in its cluster
         \item rho: the threshold for the noise cluster
          \item test_one_cluster: decision and pvalue to know if there is more than one cluster
         }
}
\description{
This function shows the clustatis results
}
\seealso{
\code{\link{clustatis}} , \code{\link{clustatis_kmeans}}
}
\keyword{quantitative}
