% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.statis.R
\name{summary.statis}
\alias{summary.statis}
\title{Show the STATIS results}
\usage{
\method{summary}{statis}(object, ...)
}
\arguments{
\item{object}{object of class 'statis'.}

\item{...}{further arguments passed to or from other methods}
}
\value{
a list with:
        \itemize{
         \item homogeneity: homogeneity of the blocks (in percentage)
         \item weights: the weights associated with the blocks to build the compromise
         \item eigenvalues: the eigenvalues of the svd decomposition
         \item inertia: the percentage of total variance explained by each axis
         }
}
\description{
This function shows the STATIS results
}
\seealso{
\code{\link{statis}}
}
\keyword{quantitative}
