% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catatis.R
\name{catatis}
\alias{catatis}
\title{Perform the CATATIS method on different blocks of binary data from a CATA experiment}
\usage{
catatis(Data,nblo,NameBlocks=NULL, NameVar=NULL, Graph=TRUE, Graph_weights=TRUE)
}
\arguments{
\item{Data}{data frame or matrix where the blocks of binary variables are merged horizontally. If you have a different format, see \code{\link{change_cata_format}}}

\item{nblo}{numerical. Number of blocks (subjects).}

\item{NameBlocks}{string vector. Name of each block (subject). Length must be equal to the number of blocks. If NULL, the names are S1,...Sm. Default: NULL}

\item{NameVar}{string vector. Name of each variable (attribute, the same names for each subject). Length must be equal to the number of attributes. If NULL, the colnames of the first block are taken. Default: NULL}

\item{Graph}{logical. Show the graphical representation? Default: TRUE}

\item{Graph_weights}{logical. Should the barplot of the weights be plotted? Default: TRUE}
}
\value{
a list with:
        \itemize{
         \item S: the S matrix: a matrix with the similarity coefficient among the subjects
         \item compromise: a matrix which is the compromise of the subjects (akin to a weighted average)
         \item weights: the weights associated with the subjects to build the compromise
         \item lambda:  the first eigenvalue of the S matrix
         \item overall error: the error for the CATATIS criterion
         \item error_by_conf: the error by subject (CATATIS criterion)
         \item s_with_compromise: the similarity coefficient of each subject with the compromise
         \item homogeneity: homogeneity of the subjects (in percentage)
         \item CA: the results of correspondance analysis performed on the compromise dataset
         \item eigenvalues: the eigenvalues associated to the correspondance analysis
         \item inertia: the percentage of total variance explained by each axis of the CA
         \item error_by_obj: the error by object (CATATIS criterion)
         \item scalefactors: the scaling factors of each subject
         \item nb_1: the number of 1 in each block, i.e. the number of checked attributes by subject.
         \item param: parameters called
         }
}
\description{
CATATIS method on binary blocks. Additional outputs are also computed
}
\examples{
data(straw)
res.cat=catatis(straw, nblo=114)
summary(res.cat)

}
\references{
Llobell, F., Cariou, V., Vigneau, E., Labenne, A., & Qannari, E. M. (2019). A new approach for the analysis of data and the clustering of subjects in a CATA experiment. Food Quality and Preference, 72, 31-39.
}
\seealso{
\code{\link{plot.catatis}}, \code{\link{summary.catatis}}, \code{\link{cluscata}}, \code{\link{change_cata_format}}
}
\keyword{CATA}
