\name{ClustBlock-package}
\alias{ClustBlock-package}
\alias{ClustBlock}
\docType{package}
\title{Clustering of Datasets}
\description{ Hierarchical and partitioning algorithms of blocks of variables.The CLUSTATIS method and the CLUSCATA method are the core of this package. The CATATIS methods allows to compute some indices and tests to control the quality of CATA data. Multivariate analysis and clustering of subjects for quantitative multiblock data, CATA, RATA, Free Sorting and JAR experiments are available.}
\details{
  \tabular{ll}{
    Package: \tab ClustBlock\cr
    Type: \tab Package\cr
    Version: \tab 3.1.0\cr
    First version Date: \tab 2019-03-06\cr
    Last version Date: \tab 2023-03-03\cr
  }
}

\author{
  Fabien Llobell, Evelyne Vigneau, Veronique Cariou, El Mostafa Qannari

  Maintainer: \email{fllobell@hotmail.fr}
}

\references{
\tabular{l}{
Llobell, F., Cariou, V., Vigneau, E., Labenne, A., & Qannari, E. M. (2020). Analysis and clustering of multiblock datasets\cr by means of the STATIS and CLUSTATIS methods. Application to sensometrics. Food\cr Quality and Preference, 79, 103520.\cr
Llobell, F., Vigneau, E., & Qannari, E. M. (2019). Clustering datasets by means of CLUSTATIS with identification\cr of atypical datasets. Application to sensometrics. Food Quality and Preference, 75, 97-104.\cr
Llobell, F., Cariou, V., Vigneau, E., Labenne, A., & Qannari, E. M. (2019). A new approach for the analysis of data \cr and the clustering of subjects in a CATA experiment. Food quality and preference, 72, 31-39.\cr
Llobell, F., Giacalone, D., Labenne, A., & Qannari, E. M. (2019). Assessment of the agreement and cluster analysis of\cr  the respondents in a CATA experiment. Food Quality and Preference, 77, 184-190.\cr
Llobell, F., & Qannari, E. M. (2020). CLUSTATIS: Cluster analysis of blocks of variables. Electronic\cr Journal of Applied Statistical Analysis, 13(2), 436-453.\cr
Llobell, F. (2020). Classification de tableaux de données, applications en analyse sensorielle (Doctoral\cr dissertation, Nantes, Ecole nationale vétérinaire).}}

\keyword{Clustering}
\keyword{Blocks}
\keyword{Check-All-That-Apply}
\keyword{Projective mapping}
\keyword{Profiling}
\keyword{Sensory}
\keyword{Noise Cluster}
\keyword{CATA}
\keyword{Cluster analysis}
\keyword{Free Sorting}
\keyword{Just About Right}
\keyword{JAR}
\keyword{RATE-All-That-Apply}
\keyword{RATA}
