\name{ClustBlock-package}
\alias{ClustBlock-package}
\alias{ClustBlock}
\docType{package}
\title{Clustering of Datasets}
\description{ Hierarchical and partitioning algorithms of blocks of variables.The CLUSTATIS method and the CLUSCATA method are the core of this package}
\details{
  \tabular{ll}{
    Package: \tab ClustBlock\cr
    Type: \tab Package\cr
    Version: \tab 2.2.0\cr
    First version Date: \tab 2019-11-29\cr
    Last version Date: \tab 2020-04-23\cr
  }
}

\author{
  Fabien Llobell, Evelyne Vigneau, Veronique Cariou, El Mostafa Qannari

  Maintainer: \email{fabien.llobell@oniris-nantes.fr}
}

\references{
\tabular{l}{
Llobell, F., Cariou, V., Vigneau, E., Labenne, A., & Qannari, E. M. (2020). Analysis and clustering of multiblock datasets by means of the STATIS and CLUSTATIS methods. Application to sensometrics. Food Quality and Preference, 79, 103520.\cr
Llobell, F., Vigneau, E., & Qannari, E. M. (2019). Clustering datasets by means of CLUSTATIS with identification of atypical datasets. Application to sensometrics. Food Quality and Preference, 75, 97-104.\cr
Llobell, F., Cariou, V., Vigneau, E., Labenne, A., & Qannari, E. M. (2019). A new approach for the analysis of data and the clustering of subjects in a CATA experiment. Food quality and preference, 72, 31-39.\cr
Llobell, F., Giacalone, D., Labenne, A., & Qannari, E. M. (2019). Assessment of the agreement and cluster analysis of the respondents in a CATA experiment. Food Quality and Preference, 77, 184-190.}}

\keyword{Clustering}
\keyword{Blocks}
\keyword{Check-All-That-Apply}
\keyword{Projective mapping}
\keyword{Sensory}
\keyword{Noise Cluster}

