\name{ClustBlock-package}
\alias{ClustBlock-package}
\alias{ClustBlock}
\docType{package}
\title{Clustering of Datasets}
\description{ Hierarchical and partitioning algorithms of blocks of variables.The CLUSTATIS method and the CLUSCATA method are the core of this package. The CATATIS methods allows to compute some indices and tests to control the quality of CATA data. Multivariate analysis and clustering of subjects for quantitative multiblock data, CATA, RATA, Free Sorting and JAR experiments are available. Clustering of rows in multi-block context (notably with ClusMB strategy) is also included.}
\details{
  \tabular{ll}{
    Package: \tab ClustBlock\cr
    Type: \tab Package\cr
    Version: \tab 4.1.0\cr
    First version Date: \tab 2019-03-06\cr
    Last version Date: \tab 2025-05-13\cr
  }
}

\author{
  Fabien Llobell, Evelyne Vigneau, Veronique Cariou, El Mostafa Qannari

  Maintainer: \email{fabienllobellresearch@gmail.com}
}

\references{
\tabular{l}{
Llobell, F., Cariou, V., Vigneau, E., Labenne, A., & Qannari, E. M. (2020). Analysis and clustering of multiblock datasets\cr by means of the STATIS and CLUSTATIS methods. Application to sensometrics. Food Quality and Preference,\cr 79, 103520.\cr
Llobell, F., Vigneau, E., & Qannari, E. M. (2019). Clustering datasets by means of CLUSTATIS with identification\cr of atypical datasets. Application to sensometrics. Food Quality and Preference, 75, 97-104.\cr
Llobell, F., Cariou, V., Vigneau, E., Labenne, A., & Qannari, E. M. (2019). A new approach for the analysis of data \cr and the clustering of subjects in a CATA experiment. Food quality and preference, 72, 31-39.\cr
Llobell, F., Giacalone, D., Labenne, A., & Qannari, E. M. (2019). Assessment of the agreement and cluster analysis of\cr  the respondents in a CATA experiment. Food Quality and Preference, 77, 184-190.\cr
Llobell, F., & Qannari, E. M. (2020). CLUSTATIS: Cluster analysis of blocks of variables. Electronic\cr Journal of Applied Statistical Analysis, 13(2), 436-453.\cr
Llobell, F. (2020). Classification de tableaux de donnees, applications en analyse sensorielle (Doctoral\cr dissertation, Nantes, Ecole nationale veterinaire).\cr
Llobell, F., Bonnet, L., & Giacalone, D. (2024). Assessment of panel performance in CATA and RATA experiment. \cr Journal of Sensory Studies, 39(4), e12941.\cr
Llobell, F., & Giacalone, D. (2025). Two Methods for Clustering Products in a Sensory Study: STATIS and ClusMB. \cr Journal of Sensory Studies, 40(1), e70024.}}

\keyword{Clustering}
\keyword{Blocks}
\keyword{Check-All-That-Apply}
\keyword{Projective mapping}
\keyword{Profiling}
\keyword{Sensory}
\keyword{Noise Cluster}
\keyword{CATA}
\keyword{Cluster analysis}
\keyword{Free Sorting}
\keyword{Just About Right}
\keyword{JAR}
\keyword{Rate-All-That-Apply}
\keyword{RATA}
\keyword{Multi-block}
