% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RATAchoc.R
\docType{data}
\name{RATAchoc}
\alias{RATAchoc}
\title{RATA data on chocolates}
\format{
RATA data with sessions. A data frame with 3 sessions, 9 panelists, 12 products and 27 RATA attributes.
}
\usage{
data(RATAchoc)
}
\description{
RATA data on chocolates
}
\examples{
data(RATAchoc)
}
\references{
Pangborn 2023
}
\keyword{datasets}
